+++ packages/contractkit/src/explorer/block-explorer.ts
@@ -1,21 +1,13 @@
-import { Address } from '@celo/utils/lib/address'
 import { Block, Transaction } from 'web3-eth'
-import abi, { ABIDefinition } from 'web3-eth-abi'
-import { CeloContract } from '../base'
+import abi from 'web3-eth-abi'
 import {
   getInitializeAbiOfImplementation,
-  PROXY_ABI,
   PROXY_SET_AND_INITIALIZE_IMPLEMENTATION_SIGNATURE,
   PROXY_SET_IMPLEMENTATION_SIGNATURE,
 } from '../governance/proxy'
 import { ContractKit } from '../kit'
 import { parseDecodedParams } from '../utils/web3-utils'
-import {
-  ContractDetails,
-  getContractDetailsFromContract,
-  mapFromPairs,
-  obtainKitContractDetails,
-} from './base'
+import { BaseExplorer } from './base'
 
 export interface CallDetails {
   contract: string
@@ -34,36 +26,15 @@ export interface ParsedBlock {
   parsedTx: ParsedTx[]
 }
 
-interface ContractMapping {
-  details: ContractDetails
-  fnMapping: Map<string, ABIDefinition>
-}
-
 export async function newBlockExplorer(kit: ContractKit) {
-  return new BlockExplorer(kit, await obtainKitContractDetails(kit))
+  const blockExplorer = new BlockExplorer(kit)
+  await blockExplorer.init()
+  return blockExplorer
 }
 
-const getContractMappingFromDetails = (cd: ContractDetails) => ({
-  details: cd,
-  fnMapping: mapFromPairs(
-    (cd.jsonInterface.concat(PROXY_ABI) as ABIDefinition[])
-      .filter((ad) => ad.type === 'function')
-      .map((ad) => [ad.signature, ad])
-  ),
-})
-
-export class BlockExplorer {
-  private addressMapping: Map<Address, ContractMapping>
-
-  constructor(private kit: ContractKit, readonly contractDetails: ContractDetails[]) {
-    this.addressMapping = mapFromPairs(
-      contractDetails.map((cd) => [cd.address, getContractMappingFromDetails(cd)])
-    )
-  }
-
-  async updateContractDetailsMapping(name: string, address: string) {
-    const cd = await getContractDetailsFromContract(this.kit, name as CeloContract, address)
-    this.addressMapping.set(cd.address, getContractMappingFromDetails(cd))
+export class BlockExplorer extends BaseExplorer {
+  constructor(kit: ContractKit) {
+    super(kit, 'function')
   }
 
   async fetchBlockByHash(blockHash: string): Promise<Block> {
@@ -120,7 +91,7 @@ export class BlockExplorer {
     const callSignature = input.slice(0, 10)
     const encodedParameters = input.slice(10)
 
-    const matchedAbi = contractMapping.fnMapping.get(callSignature)
+    const matchedAbi = contractMapping.abiMapping.get(callSignature)
     if (matchedAbi == null) {
       return null
     }
@@ -147,6 +118,7 @@ export class BlockExplorer {
         abi.decodeParameters(initializeAbi.inputs!, encodedInitializeParameters)
       )
       params[`initialize@${abi.encodeFunctionSignature(initializeAbi)}`] = initializeParams
+      delete params.callbackData
     }
 
     return {